Add Typescript for Router
Add Typescript types for the Router file.
